Basic Rendering
Screen is used to manage a buffer which can be rendered to a terminal.

At the moment it just does some simple bounds checking to prevent you
writing to areas which would be outside of the terminal.

Rendering involves creating a string which would fill the entire
terminal

At the moment the harness handles all of the terminal interactions (such
as finding the terminal size and writing the string produced by the
Screen)